Cleanup code in System.Runtime.Serialization.Formatters.Tests
