+++ src/Common/tests/System/Runtime/Serialization/Formatters/BinaryFormatterHelpers.cs
@@ -88,5 +88,42 @@ namespace System.Runtime.Serialization.Formatters.Tests
             Exception ex = Assert.Throws<TargetInvocationException>(() => constructor.Invoke(new object[] { info, new StreamingContext() }));
             Assert.IsType<PlatformNotSupportedException>(ex.InnerException);
         }
+
+        public static byte[] ToByteArray(object obj, 
+            FormatterAssemblyStyle assemblyStyle = FormatterAssemblyStyle.Full)
+        {
+            BinaryFormatter bf = new BinaryFormatter();
+            bf.AssemblyFormat = assemblyStyle;
+            using (MemoryStream ms = new MemoryStream())
+            {
+                bf.Serialize(ms, obj);
+                return ms.ToArray();
+            }
+        }
+
+        public static string ToBase64String(object obj, 
+            FormatterAssemblyStyle assemblyStyle = FormatterAssemblyStyle.Full)
+        {
+            byte[] raw = ToByteArray(obj, assemblyStyle);
+            return Convert.ToBase64String(raw);
+        }
+
+        public static object FromByteArray(byte[] raw, 
+            FormatterAssemblyStyle assemblyStyle = FormatterAssemblyStyle.Full)
+        {
+            var binaryFormatter = new BinaryFormatter();
+            binaryFormatter.AssemblyFormat = assemblyStyle;
+            using (var serializedStream = new MemoryStream(raw))
+            {
+                return binaryFormatter.Deserialize(serializedStream);
+            }
+        }
+
+        public static object FromBase64String(string base64Str, 
+            FormatterAssemblyStyle assemblyStyle = FormatterAssemblyStyle.Full)
+        {
+            byte[] raw = Convert.FromBase64String(base64Str);
+            return FromByteArray(raw, assemblyStyle);
+        }
     }
 }
